requirements.txt
fastapi
uvicorn[standard]
apscheduler
pytz
pydantic-settings
asyncpg
greenlet
bcrypt==4.0.1
cffi
passlib[bcrypt]==1.7.4
langchain-google-genai>=2.0.0,<3.0.0
langchain-text-splitters
langgraph
# sentence-transformers/faiss/torch moved to requirements-ml.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2  #
torch>=2.0.0  # 
transformers>=4.30.0  # 
python-docx>=1.1.0
google-generativeai>=0.3.2
psutil>=5.9.5
langchain-core>=0.1.17
langchain-text-splitters>=0.0.1
